% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyformatr.R
\name{easy_format}
\alias{easy_format}
\title{Easily build format strings}
\usage{
easy_format(..., sep = "")
}
\arguments{
\item{...}{A mixture of time or number components, including bases and modifications, and strings.}

\item{sep}{A separator, defaults to ""}
}
\description{
Use `filter_info` to find more information about format components
}
\examples{
easy_format(year, month, day, integer, octal, double)
easy_format(decimal(second) )
easy_format(before_decimal(double, 3) )
easy_format(month,
            roman(list(day,
                       minute) ) )
}

