% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyformatr.R
\name{easy_format_}
\alias{easy_format_}
\title{Standard evaluation version of easy_format}
\usage{
easy_format_(args, sep = "")
}
\arguments{
\item{args}{An unevaluated expression, or list of expressions}

\item{sep}{A separator, defaults to ""}
}
\description{
See title
}
\examples{
easy_format_(c("hour", "minute", "second"))
}

