% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyformatr.R
\name{filter_info}
\alias{filter_info}
\title{Query documentation}
\usage{
filter_info(filter_statement)
}
\arguments{
\item{filter_statement}{A logical filter}
}
\description{
Will return a cleaned markdown table
}
\examples{
filter_info(type == "time" & component == "base")
filter_info(type == "number" & component == "base")
filter_info(type == "time" & component == "mutant" & base == "second")
filter_info(type == "number" & component == "flag")
filter_info(type == "number" & component == "option")
}

