% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyformatr.R
\name{filter_info_}
\alias{filter_info_}
\title{Standard evaluation version of filter_info}
\usage{
filter_info_(filter_expression)
}
\arguments{
\item{filter_expression}{An unevaluated logical filter}
}
\description{
Will return a cleaned markdown table
}
\examples{
filter_info_("type == 'time' & component == 'base'")
}

