% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyformatr.R
\docType{data}
\name{format_info}
\alias{format_info}
\title{Format info}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 78 rows and 23 columns.}
\usage{
format_info
}
\description{
A huge table of documentation. Best queried through filter_info
}
\keyword{datasets}

