% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{name_value_to_list}
\alias{name_value_to_list}
\title{Convert a dataframe that has a name and a value column to a named list.}
\usage{
name_value_to_list(x)
}
\arguments{
\item{x}{A dataframe}
}
\description{
Convenience function
}

