% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{paste_NA_omit}
\alias{paste_NA_omit}
\title{Will paste, omitting all NA values}
\usage{
paste_NA_omit(..., sep = "", collapse = NULL)
}
\arguments{
\item{...}{See paste documentation}

\item{sep}{See paste documentation}

\item{collapse}{See paste documentation}
}
\description{
See title
}

