% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{paste_NA_poison}
\alias{paste_NA_poison}
\title{Will paste. Anything pasted to NA will return NA}
\usage{
paste_NA_poison(..., sep = "", collapse = NULL)
}
\arguments{
\item{...}{See paste documentation}

\item{sep}{See paste documentation}

\item{collapse}{See paste documentation}
}
\description{
See title
}

