% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{remove_NA_columns}
\alias{remove_NA_columns}
\title{Remove columns that are all NA}
\usage{
remove_NA_columns(x)
}
\arguments{
\item{x}{A dataframe or list of vectors.}
}
\description{
Convenience function
}
\examples{
remove_NA_columns(data.frame(a = NA, b = 1))
}

