% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{remove_single_value_columns}
\alias{remove_single_value_columns}
\title{Remove columns that are all identical}
\usage{
remove_single_value_columns(x)
}
\arguments{
\item{x}{A dataframe or list of vectors.}
}
\description{
Convenience function
}
\examples{
remove_single_value_columns(data.frame(a = c(1, 1), b = c(1, 2)))
}

