% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{string_table}
\alias{string_table}
\title{Build objects and dataframes together tidily}
\usage{
string_table(x)
}
\arguments{
\item{x}{A list, dataframe, string, or object that can be converted to a string.}
}
\description{
This will recursively bind dataframes together rowwise.
Anything that is not a list or a dataframe will be converted a string, doubling all percentage signs,
and then to dplyr::0data_frame(base = "raw", raw_value = .)
}

